+++ cmd/mailflow/main.go
@@ -22,7 +22,7 @@ import (
 	"github.com/mojoaar/icloud-mailflow/internal/web"
 )
 
-var version = "0.5.1"
+var version = "0.6.0"
 
 type App struct {
 	Config   *config.Config